update register and add confirm email api
