--- TicketResell_API/Controllers/User/EmailConfirmation.cs
@@ -1,8 +1,0 @@
-﻿namespace TicketResell_API.Controllers.User
-{
-    public class EmailConfirmation
-    {
-        public string? email { get; set; }
-        public int? code { get; set; }
-    }
-}
