--- TicketResell_API/Controllers/User/IEmailSender.cs
@@ -1,9 +1,0 @@
-﻿
-namespace TicketResell_API.Controllers.User
-{
-    public interface IEmailSender
-    {
-        Task<string> SendConfirmationEmailAsync(string? email, string emailCode);
-        Task SendEmailAsync(string? email, string callback, string v);
-    }
-}
