--- TicketResell_API/Controllers/User/Login.cs
@@ -1,8 +1,0 @@
-﻿namespace TicketResell_API.Controllers.User
-{
-    public class Login
-    {
-        public string email {  get; set; } = string.Empty;
-        public string password { get; set; } = string.Empty;
-    }
-}
