+++ TicketResell_API/Controllers/User/Model/EmailConfirmation.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.Build.Framework;
+using System.ComponentModel.DataAnnotations;
+
+namespace TicketResell_API.Controllers.User.Model
+{
+    public class EmailConfirmation
+    {
+
+
+        public string? email { get; set; }
+
+        public int? code { get; set; }
+    }
+}
