+++ TicketResell_API/Controllers/User/Model/Login.cs
@@ -1,0 +1,12 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace TicketResell_API.Controllers.User.Model
+{
+    public class Login
+    {
+        [Required]
+        public string email { get; set; } = string.Empty;
+        [Required]
+        public string password { get; set; } = string.Empty;
+    }
+}
