+++ TicketResell_API/Controllers/User/Model/Profile.cs
@@ -1,0 +1,12 @@
+﻿namespace TicketResell_API.Controllers.User.Model
+{
+    public class Profile
+    {
+        public string userID { get; set; }
+        public string firsName { get; set; }
+        public string lastName { get; set; }
+        public string email { get; set; }
+        public string phoneNumber { get; set; }
+
+    }
+}
