+++ TicketResell_API/Controllers/User/Model/UserRole.cs
@@ -1,0 +1,8 @@
+﻿namespace TicketResell_API.Controllers.User.Model
+{
+    public class UserRole
+    {
+        public string email { get; set; } = string.Empty;
+        public string role { get; set; } = string.Empty;
+    }
+}
