--- TicketResell_API/Controllers/User/Profile.cs
@@ -1,12 +1,0 @@
-﻿namespace TicketResell_API.Controllers.User
-{
-    public class Profile
-    {
-        public string userID {  get; set; }
-        public string firsName { get; set; } 
-        public string lastName { get; set; }
-        public string email { get; set; }
-        public string phoneNumber {  get; set; }
-
-    }
-}
