+++ TicketResell_API/Controllers/User/Service/IEmailSender.cs
@@ -1,0 +1,8 @@
+﻿namespace TicketResell_API.Controllers.User.Service
+{
+    public interface IEmailSender
+    {
+        Task<string> SendConfirmationEmailAsync(string? email, string emailCode);
+        Task SendEmailAsync(string? email, string callback, string v);
+    }
+}
