+++ TicketResell_API/Controllers/User/UpdateProfile.cs
@@ -1,0 +1,7 @@
+﻿namespace TicketResell_API.Controllers.User
+{
+    public class UpdateProfile
+    {
+        
+    }
+}
