--- TicketResell_API/Controllers/User/UserRole.cs
@@ -1,8 +1,0 @@
-﻿namespace TicketResell_API.Controllers.User
-{
-    public class UserRole
-    {
-        public string email {  get; set; } = string.Empty;
-        public string role {  get; set; } = string.Empty ;
-    }
-}
